--- src/App.css
@@ -1,35 +1,0 @@
-.QuizQuestion {
-  display: flex;
-  justify-content: center;
-  font-size: 2em;
-  text-align: center;
-}
-.QuizEnd {
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-  text-align: center;
-  font-size: 2em;
-  padding-top: 20px;
-}
-.QuizEnd a {
-  display: block;
-  padding-top: 20px;
-}
-ul {
-  padding: 0;
-  margin: 0;
-}
-.buttons {
-  display: flex;
-  justify-content: center;
-  text-align: center;
-}
-button {
-  margin: 0.6em auto;
-  font-size: 1.5em;
-  display: block;
-}
-.error {
-  color: red;
-}
